feat: mandatory 도형들 구현 완료(원기둥, 구체, 평면)
