+++ cy_test.rt
@@ -1,0 +1,9 @@
+A		 0.5				 													255,255,255
+
+C		 0,0,0		 		0,0,1.0 			90
+L		 -50,0,0		 								0.7 					255,255,255
+
+# cy	0,0,15 	1.0,1.0,1.0 		10.2 	11.42 		255,0,0
+# cy	0,0,15 	1.0,1.0,1.0 		10.2 	11.42 		255,0,0
+# cy	0,0,15 	1.0,1.0,1.0 		10.2 	11.42 		255,0,0
+cy	0,5,15 	1.0,1.0,0 		10.2 	11.42 		255,0,0
