+++ mini.rt
@@ -1,14 +1,14 @@
-A		 0.123456789				 													255,100,255
+A		 0.6789				 													255,100,255
 
 C		 0,0,0		 		0,0,1.0 			90
-L		 -40,0,30		 								0.7 					255,255,255
+L		 -40,0,40		 								0.7 					255,255,255
 
-sp	 100,0,200											200						0,0,255
-sp	 0,0,200											200						255,0,0
-pl	 0,0,0					0,1.0,0 									255,0,225
-pl	 0,0,0					0,1.0,0 									250,0,225
+sp	 50,0,200											70						0,0,255
+sp	 100,0,200											70						255,0,0
+pl	 0,0,400					0,0,1.0 									255,0,225
+pl	 0,0,400					0.7,1.0,0.5 									2,170,25
 
-cy	 50.0,0.0,20.6 	0,0,1.0 		14.2 	21.42 	10,0,255
+cy	 -50.0,10.0,70.6 	0,1.0,0 		14.2 	21.42 	255,0,10
 
-cy	 50.0,0.0,20.6 	0,0,1.0 		14.2 	21.42 	10,0,255
-L		 -20,0,40		 								0.8 					200,215,225
+cy	 -50.0,0.0,50.6 	0,0,1.0 		14.2 	21.42 	10,0,255
+L		 40,0,40		 								0.8 					200,215,225
