+++ test.rt
@@ -7,4 +7,4 @@ sp	 -100,0,200			 								140						255,0,0
 sp	 0,0,100			 								70						0,255,0
 sp	 100,0,200			 								200						0,0,255
 
-cy	100,-30,140 	1.0,1.0,1.0 		10.2 	11.42 		180,95,125
+cy	0,0,30 	1.0,1.0,1.0 		10.2 	11.42 		255,0,0
